requirements.txt
pika
boto3
PyYaml